requirement.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.140
numpy==2.1.3
openai==1.54.3
streamlit==1.39.0
transformers==4.46.2
faiss-cpu==1.9.0
altair==5.4.1